--- a/src/Microsoft.AspNetCore.SignalR.Protocols.MsgPack/Internal/Protocol/MessagePackHubProtocol.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Protocols.MsgPack/Internal/Protocol/MessagePackHubProtocol.cs
@@ -322,7 +322,9 @@ namespace Microsoft.AspNetCore.SignalR.Internal.Protocol
                 packer.PackString(message.InvocationId);
             }
             packer.PackString(message.Target);
-            packer.PackObject(message.Arguments, SerializationContext);
+            packer.PackArrayHeader(message.Arguments.Length);
+            foreach (var arg in message.Arguments)
+                packer.PackObject(arg, SerializationContext);
         }
 
         private void WriteStreamInvocationMessage(StreamInvocationMessage message, Packer packer)