@davidfowl 2018-04-04
Why didn't you make the change in general for all invocation types? (StreamInvocation etc)